server/groceryecommerce/requirements.txt
asgiref==3.8.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
dj-database-url==2.1.0
Django==5.0.3
django-cors-headers==4.3.1
django-mpesa==2.0.14
djangorestframework==3.15.1
drf-yasg==1.21.7
gunicorn==22.0.0
idna==3.7
inflection==0.5.1
packaging==24.0
pillow==10.2.0
psycopg2==2.9.9
pycparser==2.22
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.10.0
uritemplate==4.1.1
urllib3==2.2.1
whitenoise==6.6.0
